Issue 11: Switch from unsigned long to size_t in HyperFragLookupTable::split() function.

+++ comet-ms/CometSearch/PeptideFragmentation/MakeMS2/FragmentModels/HyperFragLookupTable.cpp
@@ -13,8 +13,8 @@ using namespace std;
 
 void HyperFragLookupTable::split(const string& s, string delimiter, vector<string>& v) {
 
-    unsigned long posPrevious = 0;
-    unsigned long posCurrent = s.find(delimiter, posPrevious);
+    size_t posPrevious = 0;
+    size_t posCurrent = s.find(delimiter, posPrevious);
 
     while (posCurrent != string::npos) {
